Bugfixes in lectures 7, 10
For lecture 7, there was a typo which is fixed now.
The corrected version now say:

Similarly both LowQualFinSF and BsmtUnfSF have a negative
impact on the prediction but BsmtUnfSF would have a bigger
negative impact because it's going to be multiplied by a
number with a bigger magnitude.

For lecture 10, the following line is commented out:
from pandas_profiling import ProfileReport

This helps to run the lecture notebook even if ProfileReport
has not be installed

There has been NO conceptual changes made for these lecture notes.